requirements.txt
django==2.0.2
djangorestframework==3.10.3
django-configurations==2.2
tweepy==3.8.0
tweet-preprocessor==0.5.0
google-cloud-language==1.3.0
nltk==3.3